First draft of the GitHub integration
This just adds an endpoint that listens to the `push` webhook (set up
manually on GitHub). Right now it clones the whole repo in a temp dir
gets all the file paths and chooses the ones that look like tables.
With this we create a payload that is passed to the goodtables celery
task.

Lots of stuff to improve, but it works! Ie files get validated on each
push