Save job statuses on the DB
Jobs are always created with a status of `created`. Plugins or the
validation task will set it to `running` when the job starts. After
goodtables runs it will be set to `success` or `failure` depending on
the report.

TODO: set to `error` on exceptions

Also improve the way db connections are handled