Switch to SQLAlchemy ORM Models
Drop dataset and use pure SQLAlchemy models.

I'm still not 100% happy with how the tasks don't use the same interface
for getting (and updating) jobs than the rest of the app, just because
they don't use the same session. We can probably pass the session as an
optional parameter to the helpers and use these consistently.