Refactor so all model calls are done in helpers
Rather than having SQLAlchemy all over the place let's keep these at the
helpers level so plugins or tasks don't need to worry about SA calls.

This implies that sometimes you need to explicitly pass a session if the
one you are using is not the default one one `goodtablesio.services` (ie
you are in a task). I think the benefits of conceptually separating the
interaction with the DB outweigh this minor incovenience.

Improved helpers docstrings.