New model for plans and subscriptions
New model class and table for Plans. Each user has a Subscription to one
of the plans. There is a `User.set_plan` method that creates / updates
the subscription, and another method for extending it, but the exact
details on how this will work will be defined once we know more about
the Stripe integration (#118)