Handle private repos on account sync
Do not link users to private repos by default (ie if they are on the
free plan or they are not admins).

This needs a much more comprehensive handling, but for the first launch
without plans should be enough. Main TODO: when finding a private repo,
check that there is an active subcription (from the current user or
another one)  that contains the GitHub organization for this particular
repo.

This will require that subcriptions are linked to a particular Github
org(s).