Move NodeJS-related scripts from Makefile to package.json
The Makefile then is simply a dumb wrapper that calls commands defined in the
`package.json` or `tox.ini`, depending if it's JavaScript or Python. This
allows us to rely on their respective package manager to run the correct
version of each executable (avoiding having to mess with $PATH).

I also changed the testing NODE_ENV to "test", as it's the most common name.