WIP Fixing Docker build
Webpack was failing with a segfault when building the container, which was
fixed by upgrading NodeJS (the old version was 6.10). I couldn't install it via
Alpine, so I ended up upgrading the base container we use to Python 3.6 and
Alpine 3.7, which has a NodeJS 8.x package.

This upgrade also broke the postgresql package, because it's compiled against
libressl, and not openssl. I switched to libressl, and had to also upgrade the
`cryptography` python library to a newer version that supports libressl.